my-app/src/backend/requirements.txt
lxml==4.9.1
python-dateutil==2.8.2